Generalized get_corpus.py and get_vm_tid.py to handle multiple threads (i.e., VMs)
 Changes to be committed:
	deleted:    document_clustering.py
	modified:   get_corpus.py
	modified:   get_vm_tid.py
	modified:   trace_clustering.py

get_corpus.py:
Now handles multiple tids and extracts both events and their timestamps in the form of a dict indexed by tid values of qemu running the VMs.
Note that the layout of the trace path .tid file is changed to contain multiple tids in the form of:

N:trace-path-filename:tid1:tid2:tid3

where N is the class assigned to this trace file

get_vm_tid.py:
Now can extract multiple tids each belonging to an independent VM that is run by a qemu thread
The output .tid file is formatted according to above

trace_clustering.py: now stores the top most frequent patterns in a file as significant patterns. This file is loadable for later consideration.
However, the part where only the most significant patterns are considered for tokens is not yet implemented. Instead we use the --n-features option.
In practice, however, we would need to limit this to a static list of patterns for speed purposes. This can potentially reduce the resource footprint
of the clustering algorithm.