Multi VM capability fully built in  Changes to be committed: 	modified:   trace_clustering.py
New .tid file syntax to be used:
Example:
-----------------------
5,5,5,5:./extra/traces/apache-100000-delay-100-20-25-normal/kernel:20257:20256:20255:20245
3:/home/azhari/linux_tracing/traces/mtx-20171228-205202/kernel:7202
4:/home/azhari/linux_tracing/traces/file-20171228-204640/kernel:7202
2:/home/azhari/linux_tracing/traces/cpu-20171228-204259/kernel:7202
------------------------
First line means there are four tids and all are assigned a class number of 5
Last line is commented and will not be considered for clustering but will be
considered for prediction.

General format of .tid files:
class1,class2,class3,...:trace-path:tid1,tid2,tid3,...
a # sign at the begining of a line will exclude it from clustering and include
it in prediction.